requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autopep8==1.4.2
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1.2
django-allauth==0.38.0
django-autoslug-iplweb==1.9.4
django-cors-headers==2.4.0
django-extensions==2.1.3
django-heroku==0.3.1
django-rest-swagger==2.1.0
django-versatileimagefield==1.10
djangorestframework==3.9.0
flake8==3.6.0
flake8-polyfill==1.0.2
factory-boy==2.11.1
Faker==0.9.2
gunicorn==19.9.0
idna==2.7
isort==4.3.4
itypes==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
mando==0.6.4
MarkupSafe==1.0
mccabe==0.6.1
more-itertools==4.3.0
oauthlib==2.1.0
openapi-codec==1.3.2
Pillow==5.3.0
pluggy==0.8.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
py==1.7.0
pycodestyle==2.4.0
pycparser==2.19
pyflakes==2.0.0
PyJWT==1.6.4
pylint==2.1.1
pytest==3.9.3
python-dateutil==2.7.5
python-http-client==3.1.0
python3-openid==3.1.0
pytz==2018.5
radon==2.4.0
requests==2.20.0
requests-oauthlib==1.0.0
simplejson==3.16.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24
whitenoise==4.1
wrapt==1.10.11
